remove compiler warnings about enum compare

+++ src/lpm_error_impl.hpp
@@ -123,7 +123,8 @@ void compute_error(const V1 err, const V2 appx, const V3 exact) {
 template <typename V1, typename V2, typename V3>
 void compute_error(const V1 err, const V2 appx, const V3 exact,
                    const mask_view_type m) {
-  static_assert(V1::Rank == V2::Rank and V2::Rank == V3::Rank,
+  static_assert(static_cast<int>(V1::Rank) == static_cast<int>(V2::Rank)
+            and static_cast<int>(V2::Rank) == static_cast<int>(V3::Rank),
                 "view ranks must match");
   LPM_REQUIRE(err.extent(0) == appx.extent(0) and
               appx.extent(0) == exact.extent(0));
@@ -138,7 +139,8 @@ template <typename V1, typename V2>
 ENormScalar reduce_error(const V1 err, const V2 exact,
                          const scalar_view_type wt) {
   ENormScalar rval;
-  static_assert(V1::Rank == V2::Rank, "view ranks must match");
+  static_assert(static_cast<int>(V1::Rank) == static_cast<int>(V2::Rank),
+      "view ranks must match");
   LPM_REQUIRE(err.extent(0) == exact.extent(0) and
               exact.extent(0) == wt.extent(0));
   LPM_REQUIRE(err.extent(1) == exact.extent(1));
